Make "play all" functionality baseline functional
In the file explorer make it so that the "play all" option actually
composes a playlist and plays it! Small refactor so that the audio
extension detection functionality in the player backend can be shared
among all of the compilation units.

Additional fixes for the README and forcing the default volume to 50%
instead of the 70% default in the ADF