Add tracing, move tasks to core 1, terminate decoders when switching, increase buffers
When switching between decoders we now terminate the old decoders,
deleting the task and freeing the associated memory.

There was an issue where sometimes the headphone task was getting
starved for samples leading to skips in the music - double the
output buffers for all decoders to keep this from happening, though
more investigation into why this is happening is necessary.

Add in runtime analysis functions which compile in when the associated
FreeRTOS functionality is enabled.

Share the load of some tasks by migrating main and the ADF delegation
task from core0 to core1